+++ lib/truffle/socket/truffle/foreign/sockaddr.rb
@@ -28,14 +28,28 @@ module Truffle
   module Socket
     module Foreign
       class Sockaddr < Truffle::FFI::Struct
-        config('platform.sockaddr', :sa_data, :sa_family)
+        config('platform.sockaddr', :sa_family, :sa_data)
 
-        def data
-          self[:sa_data]
+        SA_FAMILY_T = Truffle::Config['platform.typedef.sa_family_t'].to_sym
+        SA_FAMILY_OFFSET = Truffle::Config['platform.sockaddr.sa_family.offset']
+
+        def self.family_of_string(str)
+          case SA_FAMILY_T
+          when :ushort
+            str[SA_FAMILY_OFFSET..-1].unpack1('S')
+          when :uchar
+            str[SA_FAMILY_OFFSET..-1].unpack1('C')
+          else
+            raise "Unexpected type for sa_family_t: #{SA_FAMILY_T}"
+          end
         end
 
         def family
           self[:sa_family]
+        end
+
+        def data
+          self[:sa_data]
         end
 
         def to_s